mips-side: start integrating Wonderful Toolchain
Wonderful Toolchain is a collection of toolchains for several
platforms, notably including the BlocksDS project. Interestingly, it
also provides a MIPS toolchain, complete with picolibc, as well as a
few different useful libraries.

Since the DSTWO is inherently tied to the Nintendo DS, it makes sense
to reuse as much resources as BlocksDS does.

Things that happen in this commit:
- Delete zlib
  - Wonderful Toolchain has a zlib library we can use instead
- Delete libc
  - Wonderful Toolchain has picolibc integrated
- Add a cleaned out Makefile
  - Heavily based on the design of BlocksDS
- Move libds2 to the root of mips-side
  - This is needed because the include folder is needed for it to
    build, and generally easier to maintain this way.
  - It should be looked into moving both the include and libds2 into
    its own directory altogether.
- Reorganize CRT files
- Temporarily drop unpacker and unfilter examples from the base install
  script as they do not compile in this new layout